feat(genesis): expose genesis generator
