+++ src/lib.rs
@@ -27,10 +27,12 @@
 
 mod account;
 mod actor;
+mod genesis;
 mod replica;
 
 pub use self::{
-    account::Account, actor::Actor as TransferActor, replica::Replica as TransferReplica,
+    account::Account, actor::Actor as TransferActor, genesis::get_genesis,
+    replica::Replica as TransferReplica,
 };
 
 use safe_nd::{
@@ -175,18 +177,15 @@ pub struct TransferRegistrationSent {
 #[allow(unused)]
 mod test {
     use crate::{
-        actor::Actor, replica::Replica, Account, ActorEvent, ReplicaEvent, ReplicaValidator,
-        TransferInitiated,
+        actor::Actor, genesis, replica::Replica, Account, ActorEvent, ReplicaEvent,
+        ReplicaValidator, TransferInitiated,
     };
     use crdts::{
         quickcheck::{quickcheck, TestResult},
         Dot,
     };
-    use rand::Rng;
-    use safe_nd::{
-        AccountId, DebitAgreementProof, Keypair, Money, PublicKey, Result, SignedTransfer, Transfer,
-    };
-    use std::collections::{BTreeMap, HashMap, HashSet};
+    use safe_nd::{AccountId, DebitAgreementProof, Keypair, Money, PublicKey, Result, Transfer};
+    use std::collections::{HashMap, HashSet};
     use threshold_crypto::{PublicKeySet, SecretKey, SecretKeySet, SecretKeyShare};
 
     macro_rules! hashmap {
@@ -240,7 +239,7 @@ mod test {
     #[test]
     fn genesis_can_only_be_the_first() -> Result<()> {
         let debit_proof = get_genesis();
-        let mut account_configs = hashmap![0 => 10];
+        let account_configs = hashmap![0 => 10];
         let mut groups = get_network(1, 3, account_configs).0;
         let previous_key = Some(PublicKey::Bls(debit_proof.replica_keys().public_key()));
         for replica in &mut groups.remove(0).replicas {
@@ -297,7 +296,7 @@ mod test {
 
         // --- Arrange ---
         let recipient_final = sender_balance + recipient_balance;
-        let mut account_configs =
+        let account_configs =
             hashmap![sender_index => sender_balance, recipient_index => recipient_balance];
         let (_, mut actors) = get_network(group_count, replica_count, account_configs);
         let mut sender = actors.remove(&sender_index).unwrap();
@@ -342,7 +341,7 @@ mod test {
 
         sender
             .actor
-            .apply(ActorEvent::TransferInitiated(transfer.clone()));
+            .apply(ActorEvent::TransferInitiated(transfer.clone()))?;
 
         Ok(transfer)
     }
@@ -415,59 +414,13 @@ mod test {
     // ------------------------------------------------------------------------
 
     fn get_genesis() -> DebitAgreementProof {
+        let balance = u32::MAX as u64 * 1_000_000_000;
         let mut rng = rand::thread_rng();
-        let index = 0;
         let threshold = 0;
         let bls_secret_key = SecretKeySet::random(threshold, &mut rng);
         let peer_replicas = bls_secret_key.public_keys();
-        let secret_key = bls_secret_key.secret_key_share(index);
-        let mut account = Account::new(PublicKey::Bls(peer_replicas.public_key()));
-        let replica = Replica::from_snapshot(
-            secret_key.clone(),
-            index,
-            peer_replicas.clone(),
-            Default::default(),
-            Default::default(),
-            Default::default(),
-        );
-        let transfer = Transfer {
-            amount: Money::from_nano(u32::MAX as u64 * 1_000_000_000),
-            id: Dot::new(get_random_pk(), 0),
-            to: account.id(),
-        };
-
-        let serialised_transfer = bincode::serialize(&transfer).unwrap();
-        let transfer_sig_share = secret_key.sign(serialised_transfer);
-        let mut transfer_sig_shares = BTreeMap::new();
-        let _ = transfer_sig_shares.insert(0, transfer_sig_share);
-        // Combine shares to produce the main signature.
-        let actor_signature = safe_nd::Signature::Bls(
-            peer_replicas
-                .combine_signatures(&transfer_sig_shares)
-                .expect("not enough shares"),
-        );
-
-        let signed_transfer = SignedTransfer {
-            transfer,
-            actor_signature,
-        };
-
-        let serialised_transfer = bincode::serialize(&signed_transfer).unwrap();
-        let transfer_sig_share = secret_key.sign(serialised_transfer);
-        let mut transfer_sig_shares = BTreeMap::new();
-        let _ = transfer_sig_shares.insert(0, transfer_sig_share);
-        // Combine shares to produce the main signature.
-        let debiting_replicas_sig = safe_nd::Signature::Bls(
-            peer_replicas
-                .combine_signatures(&transfer_sig_shares)
-                .expect("not enough shares"),
-        );
-
-        DebitAgreementProof {
-            signed_transfer,
-            debiting_replicas_sig,
-            replica_key: peer_replicas,
-        }
+        let id = PublicKey::Bls(peer_replicas.public_key());
+        genesis::get_genesis(balance, id)
     }
 
     fn get_network(
@@ -517,7 +470,7 @@ mod test {
             to,
             amount,
         };
-        account.append(transfer);
+        let _ = account.append(transfer);
 
         TestAccount {
             account,
@@ -630,7 +583,7 @@ mod test {
     struct Validator {}
 
     impl ReplicaValidator for Validator {
-        fn is_valid(&self, replica_group: PublicKey) -> bool {
+        fn is_valid(&self, _replica_group: PublicKey) -> bool {
             true
         }
     }
